requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
transformers
accelerate
einops
fastapi
uvicorn
pydantic
requests
pillow
python-multipart